snake: moved network layer primitives to the root package, adapted the sandbox.
